removed testing files, added comments, and began working on user authentication.
Completely removed testing files and folders to begin working on the actual code for the application. I started off with implementing the basic login functionality (no styling and database connection as of yet), which allows the user to login with their email and password (created upon registration but yet to be implemented). Also did the register model which still needs proper functionality and a page as with the case of the login. Also created ApplicationUser which will be an addon to the IdentityUser for ASP.NET.