added page and functionality for user registration, updated login functionality and connected both to a table in postgresql database
Created a basic page for login and register, but enough to test the functionality. When registering, the user is required to fill in their full name, email, and password (which for now just has to be 6 characters long), and optionally their institute. After filling in the details, the user may press the register button, and it should add their credentials to the AspNetUsers table (default for Asp.Net.Identity) where they are also given a unique id (jwt token) and their password is automatically hashed. Then when the user wants to login, they simple type "/login", and fill in the email and password they used to register.