Created Fish Species table and updated AppDBContext to include DbSet and table mapping.
Created the Fish Species table with columns defined in backend/models/Species.cs and generates the corresponding migration files. Additionally, it updates the AppDBContext.cs file to include a DbSet<SpeciesModel> for the Species table and configures the table mapping using the OnModelCreating method. The data that will be inserted into the table will be made available by the APIs.